Allow timer to handle user inputted values
- Add lots of functionality to app.js. Allow for a user to input time
in minutes. Values are checked and validated by: 1) not allowing the
user to enter non-numbers through the UI; 2) by checking the inputted
values and returning "reasonable" defaults. Add more functionality
for handling UI states too. Add pause button and rethink logic. Still
a huge work in progress, but it's at a point where I want to commit
because it's "stable" and at a state ready for more features
- Add pause icon to sprites.svg
- Lots of UI improvements in _app.scss
- Add styles for selection and hidden and clickable classes in
_general.scss
- Update footer styles to accommodate new shadow on timer
- Add shadow color variable to _variables.scss