Add session toggle and allow time resets
- Kitty Giraudel's article "An accessible toggle" was super helpful
and the basis for the current toggle implementation:
https://kittygiraudel.com/2021/04/05/an-accessible-toggle/
Will need to work on accessibility in the future
- Markup refinements in index.pug to accommodate the toggle and reset
button and their styles
- Lots of logic additions and refinements in app.js. Accommodate the
ability to toggle between sessions. Store temp and default times for
work and break session. If a custom time is entered for a session,
the value carries over, allowing the user to switch between sessions
and have their custom times saved. These are called temporary times
and can be reset to the default time (which is hardcoded currently). I
think this is at a good "save" state for the next set of features.
Automatic switching of sessions after a session is completed and local
storage are on the horizon!